+++ db/version_set.cc
@@ -1310,6 +1310,14 @@ void VersionStorageInfo::ComputeCompactionScore(
       } else {
         score = static_cast<double>(num_sorted_runs) /
                 mutable_cf_options.level0_file_num_compaction_trigger;
+        if (compaction_style_ == kCompactionStyleLevel && num_levels() > 1) {
+          // Level-based involves L0->L0 compactions that can lead to oversized
+          // L0 files. Take into account size as well to avoid later giant
+          // compactions to the base level.
+          uint64_t base_level_max_bytes = MaxBytesForLevel(base_level());
+          score = std::max(
+              score, static_cast<double>(total_size) / base_level_max_bytes);
+        }
       }
     } else {
       // Compute the ratio of current size to size limit.